requirements.txt
# dependencies used in all environments

# Non-Python dependencies
# sudo apt-get install imagemagick libexempi3 libxml2-dev python-dev pmount udisks
# Elasticsearch 2.4.3

                          # LICENSE      Python3  PyPy

beautifulsoup4==4.5.3     # MIT          y
click==6.7                # BSD          y        y
elasticsearch>=2.0.0,<3.0.0     # Apache y        y
elasticsearch-dsl>=2.0.0,<3.0.0 # Apache y        y
envoy==0.0.3              # MIT          y             TODO replace!
GitPython==1.0.0          # BSD          y
graphviz==0.6             # MIT          y
Jinja2==2.9.5             # BSD          y
markupsafe                # BSD          y             required by jinja2 but absent
psutil==5.2.1             # BSD          y        y
python-dateutil==2.6.0    # BSD          y
python-xmp-toolkit==2.0.1 # New BSD      y
pytz==2016.10             # MIT          y
requests==2.13.0          # Apache       y        y
simplejson                # MIT          y        y
unicodecsv==0.14.1        # BSD          y        y

bpython
coverage
pytest
pytest-cov
pytest-testmon
pytest-xdist
tox
